Rename new module cause we're rife with ophidiophobia.
